+++ src/TaskQueue/task.py
@@ -30,9 +30,20 @@ class Task[**P, R]:
         return await self.func(*args, **kwargs)
 
     async def submit(
-        self, group_id: str | None, /, *args: P.args, **kwargs: P.kwargs
+        self,
+        group_id: str | None,
+        job_id: str | None,
+        /,
+        *args: P.args,
+        **kwargs: P.kwargs,
     ) -> JobHandle[R]:
-        job = Job(task_name=self.name, args=args, kwargs=kwargs, group_id=group_id)
+        job = Job(
+            id=job_id,
+            task_name=self.name,
+            args=args,
+            kwargs=kwargs,
+            group_id=group_id,
+        )
         await self._backend.enqueue(job.id, job.to_record(self._serializer))
         logger.debug(
             "submitted job %s (task=%s, group=%s)", job.id, self.name, group_id